Assembly and Lan Connection Setup of 50 Desktop Computers
The purpose of this project was and set up 50 desktop computers, including the installation of the operating system (OS) and the establishment of a local area network (LAN) connection for all 50 machines. This report provides a detailed overview of project, including the equipment used, the assembly process, the installation, and the LAN setup.